with cursor pro

+++ contracts/structs/Delta.sol
@@ -2,6 +2,6 @@
 pragma solidity ^0.8.30;  
 
 struct Delta {
-    int256 asset0;
-    int256 asset1;
+    int128 asset0;
+    int128 asset1;
 }
